Use new combined task editor for task editing/creation
